Separate per-transport tests.
